README and JavaScript updates.

+++ script.js
@@ -11,7 +11,6 @@ let firstNumber;
 let secondNumber;
 let storedOperator = "";
 
-// Event listener for numbers and decimal point
 const buttons = document.querySelectorAll(".inputs");
 for (const button of buttons) {
   button.addEventListener("click", () => {
@@ -24,7 +23,6 @@ for (const button of buttons) {
   });
 }
 
-// Event listener for the operator
 const operators = document.querySelectorAll(".operator");
 for (const operator of operators) {
   operator.addEventListener("click", () => {
@@ -55,28 +53,29 @@ function operatorChosen(operator) {
 
 function performCalculation() {
   console.log(`** performCalculation : fired **`);
-  // take the first number and the operator and calculate it with the second number chosen.
   let calc = 0;
   secondNumber = document.getElementById("display").innerText;
+  // Used for debugging
   console.log(`performCalculation firstNumber : ${firstNumber}`);
   console.log(`performCalculation secondNumber : ${secondNumber}`);
   console.log(`performCalculation : ${storedOperator}`);
 
   switch (storedOperator) {
     case "/":
-      calc = parseInt(firstNumber) / parseInt(secondNumber);
+      calc = parseFloat(firstNumber) / parseFloat(secondNumber);
       break;
     case "*":
-      calc = parseInt(firstNumber) * parseInt(secondNumber);
+      calc = parseFloat(firstNumber) * parseFloat(secondNumber);
       break;
     case "-":
-      calc = parseInt(firstNumber) - parseInt(secondNumber);
+      calc = parseFloat(firstNumber) - parseFloat(secondNumber);
       break;
     case "+":
-      calc = parseInt(firstNumber) + parseInt(secondNumber);
+      calc = parseFloat(firstNumber) + parseFloat(secondNumber);
       break;
   }
   document.getElementById("display").innerText = calc.toLocaleString();
+  // Used for debugging
   console.log(`calc : ${calc}`);
   console.log(`calc typeof : ${typeof calc}`);
   return;
@@ -87,5 +86,3 @@ function clearDisplay() {
   document.getElementById("display").innerText = currentDisplay;
   return;
 }
-
-// operator function to capture the first number before applying the operator and clearing the screen
